Remove Java SE 6 from feature list.
When running with ibm.javaVersion=true the feature list reports that features run
on Java SE 6 when Liberty won't run on Java SE 6. Update to stop claiming Java SE 6.
Also it seems this option failed when run on OpenLiberty because the Java EE 6 features
are not there, so update the code to cope.